+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Replace duplicates in middleware with object wrapper calls. [PR#123](https://github.com/JsonMapper/JsonMapper/pull/123)
 - Correct code style issues. [PR#124](https://github.com/JsonMapper/JsonMapper/pull/124)
+- Return empty array for union type with an array type when value is an empty array. [PR#125](https://github.com/JsonMapper/JsonMapper/pull/125) Thanks to [template-provider](https://github.com/template-provider) for reporting the issue
 
 ## [2.9.1] - 2021-11-12
 ### Fixed